[HER-11]Follow-up-Rename the style files
