+++ src/home/carousel/styles/carousel.js
