+++ src/home/carousel/styles/content.js
